+++ CHANGELOG.md
@@ -6,4 +6,11 @@ CHANGELOG.md
 + version 1.4.11 - added more event modifiers
 + version 1.4.12 - fixed pluginsMap pawaAttribute to pawa-ssr 1.3.3
 + version 1.4.13 - 1.4.16 fixed issues for resuming component 
-+ version 1.4.18  fixed issue with component inside template and added event modifiers
++ version 1.4.18 - 1.4.19  fixed issue with component inside template and added event modifiers
++ version 1.4.19 - 1.4.21 made runEffect Global
++ version 1.4.22 fixed merger if when the ssr render else and else render down in pawajs engine
++ version 1.4.23 fixed _context from production
++ version 1.4.24 - add HMR mapping
++ version 1.4.26 - fixed key getcomment id
++ version 1.4.27 - fixed __pawaDev Tools,
++ version 1.4.28 - fixed pawaElement checkStaticsContext